Initial task system implementation, some interactable updates
Adds a robust Task system for creating Tasks that the player must complete.

Some functionality is limited, such as saving and loading Task state. No TaskManager is currently added to the scene, so Tasks are not actually active within the scene. But all the code is there to facilitate it later.

Adds an Events system, as well, since Events are the main way that different Task states are triggered.