Removed one side of the copper sulfate label
I've determined that the issue with the way the copper sulfate jar shows the text label from both sides of the object has something to do with the faces of the jar's UVs, and therefore is something I cannot quickly fix at this moment. Since time is short and there's more important stuff to get done at this time, this edit will suffice.