Various Fixes
Soap bottle in WebGL wasn't appearing as green for me. Not sure if anyone else had this issue. Fixed the name of the 250 beaker to match in both VR and WebGL scenes so that the glassware inspect can actually work for both. Added code to the wash hands step to activate the wash hands trigger so that it stops throwing errors when the player uses the sink prior to reaching that step.