Updated Collision Layers and Fixed pipette bulbs
For some reason whatever I had set on the pipettes for the bulbs didn't work anymore, and after looking at the code I'm not sure how it ever worked in the first place. Also, all the main objects in the VR scene have had their collisions edited so that they are all on the Holdable layer, and ignore other objects on the Holdable layer and the "Chem" layer (objects were colliding with the chem colliders used by the pour system so we have to ignore those too). As far as I can tell pouring still works, and many things no longer collide with each other, while all still being functional. Also updated the text in Glassware Use Overview to match the right and left trigger buttons.