Resolved Some Scene Issues
The audio play for the gloves was failing because the audio on all the PPE was not properly unsubscribing from events in the previous scene, and so certain calls in the Glassware Use Scene were returning null reference exceptions. The weigh boat was also having issues from the collision layer changes we made so I remedied those as well. Need to continue running through the glassware use scene after the tutorial to find more potential bugs and errors.